TeleChurnAI/TeleChurnAI/requirements.txt
django 
pandas 
numpy 
scikit-learn 
joblib 
djangorestframework 
